signalfx-tracing/signalfx-python-tracing/django/requirements.txt
django
djangorestframework
coreapi
pyyaml
requests
signalfx-tracing>=0.0.8